requirements.txt
feedparser==6.0.11
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
APScheduler==3.10.4
vaderSentiment==3.3.2
streamlit==1.37.1